The pre-commit hook fails indicating changes in all dependencies (there is none) when the main repo contains untracked files and committing with the "Local Changes" option in SmartGit.
Workaround for what seems like a git bug (or some incomprehension on my part)
When providing a pathspec pointing to an untracked file when committing the main
repo and the pre-commit hook looks like so: 'git -C "../some-dependency" status --porcelain'
then the output of "git status" shows all unmodified files as both
D[eleted] and ??[untracked] simultaneously. They should simply not be listed as modified in any way.